PyStationB/requirements.txt
azure-cosmosdb-table==1.0.6
azure-storage-blob==12.8.0
azureml-core==1.29.0
azureml-pipeline==1.29.0
cffi==1.14.5
cython==0.29.23
dacite==1.6.0
emcee==3.0.2
flask==2.0.1
gpy==1.10.0
matplotlib==3.4.2
munch==2.5.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
param==1.10.1
parse==1.19.0
Pillow==8.2.0
Pint==0.17
plotly==4.14.3
psycopg2-binary==2.8.6
pydoe==0.3.8
pydantic==1.8.2
python-dotenv==0.17.1
pyyaml==5.4.1
rdflib==5.0.0
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
sobol-seq==0.2.0
sqlalchemy==1.4.15
